Kviz_Hra/requirements.txt
Pillow>=9.0.0
python-dateutil>=2.8.2
